+++ lib/cart/more.dart
@@ -1,0 +1,38 @@
+import 'package:flutter/material.dart';
+
+class More extends StatelessWidget {
+  const More({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return GestureDetector(
+      onTap: () {},
+      child: Container(
+        decoration: decoration(),
+        height: 50,
+        child: Row(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            Icon(Icons.add),
+            Text(
+              '더 담으러 가기',
+              style: TextStyle(fontSize: 17),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+
+  BoxDecoration decoration() {
+    return BoxDecoration(
+      color: Colors.white,
+      border: Border(
+        bottom: BorderSide(
+          color: Colors.grey.withOpacity(0.3),
+          width: 2,
+        ),
+      ),
+    );
+  }
+}
